kernel: move internal headers into include/kernel
Move 2 headers that are internal to the kernel into include/kernel.
